requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
flask>=2.0.0
flask-cors>=3.0.0
librosa>=0.10.0
transformers>=4.30.0
soundfile>=0.12.0
cryptography>=41.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
